refactor editor-autosuggest to add return a list of completers
