+++ src/plugins/editor-autosuggest-refs/wrap-actions.js
@@ -1,19 +1,11 @@
-import isArray from "lodash/isArray"
-import getRefsForPath from "./get-refs-for-path"
+import getCompletions from "./get-completions"
 
-export const getCompletions = (ori, system) => (...args) => {
-  ori(...args)
-  // (ctx, editor, session, pos, prefix, cb)
-  const [ctx, editor,, pos, prefix, cb] = args
-  const { fn: { getPathForPosition } } = system
-  const { AST } = ctx
-  const editorValue = editor.getValue()
-  const currentLine = editorValue.split("\n")[pos.row]
-
-  const path = getPathForPosition({ pos, prefix, editorValue, AST})
-  const suggestions = getRefsForPath({ system, path, prefix, currentLine})
-
-  if(isArray(suggestions) && suggestions.length) {
-    cb(null, suggestions)
-  }
+// Add an autosuggest completer
+export const addAutosuggestionCompleters = (ori, system) => (context) => {
+  return ori(context).concat([{
+    getCompletions(...args) {
+      // Add `context`, then `system` as the last args
+      return getCompletions(...args, context, system)
+    }
+  }])
 }
