+++ src/plugins/editor-autosuggest-snippets/wrap-actions.js
@@ -1,19 +1,11 @@
-import isArray from "lodash/isArray"
-import snippets from "./snippets"
-import getSnippetsForPath from "./get-snippets-for-path"
+import getCompletions from "./get-completions"
 
-export const getCompletions = (ori, sys) => (...args) => {
-  ori(...args)
-  // (ctx, editor, session, pos, prefix, cb)
-  const [ctx, editor,, pos, prefix, cb] = args
-  const { fn: { getPathForPosition } } = sys
-  const { AST } = ctx
-  const editorValue = editor.getValue()
-
-  const path = getPathForPosition({ pos, prefix, editorValue, AST})
-  const suggestions = getSnippetsForPath({ path, snippets})
-
-  if(isArray(suggestions) && suggestions.length) {
-    cb(null, suggestions)
-  }
+// Add an autosuggest completer
+export const addAutosuggestionCompleters = (ori, system) => (context) => {
+  return ori(context).concat([{
+    getCompletions(...args) {
+      // Add `context`, then `system` as the last args
+      return getCompletions(...args, context, system)
+    }
+  }])
 }
